Fat: Fix nbgl_drawText for data larger than 255
'@' character compressed length was bigger than 255. The stored byte
count on `nbgl_drawText` was on 8 bits only, making a badly rendered '@'.
This variable is now an `uint16_t`.

(cherry picked from commit 5fc3edd1321a1eeb0c208ee370a5f1ee6622d2ba)

+++ lib_nbgl/src/nbgl_draw.c
@@ -417,7 +417,7 @@ void nbgl_drawText(const nbgl_area_t *area, const char* text, uint16_t textLen, 
     uint8_t char_y_min = 0;
     uint8_t char_x_max = 0;
     uint8_t char_y_max = 0;
-    uint8_t char_byte_cnt = 0;
+    uint16_t char_byte_cnt = 0;
     uint8_t encoding = 0;
 
     unicode = nbgl_popUnicodeChar((const uint8_t **)&text, &textLen, &is_unicode);
